requirements.txt
Django==2.0.7
django-daterange-filter==1.3.0
django-storages==1.7.1
dropbox==9.3.0
